Rendering with z-buffers
This adds rendering triangles with a Z buffer.

Adding a Z buffer meant that I wanted to use an Image full of f32s, so I
made Image generic over its pixel type.

This also fixes two bugs in BMP export:
1. The offset field was wrong, which caused colors to get distorted, or
parts of the image to be offset.
2. BMP uses BGR, but the colors were presented as being RGB. Now colors
are stored as BGR but have a Color::rgb() constructor because it's more
familiar to write colors that way.