Improve bmp so it can read upside-down images
The image viewer I use (feh) doesn't support upside-down images, but
ImageMagick outputs them when converting, so I decided to flip them here
in the library so that I could look at them. Also, we need to be able to
open the files that ImageMagick outputs here since I have lots of
textures in other formats that I want to read.