Fix a bug in loading bmps with larger headers
I was ignoring the image start offset field because I was only loading
files that I had generated previously, but ImageMagick actually produces
a larger header. Now the Read that you pass to must also be Seek so that
we can skip to the part of the image where the actual image data begins.